+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -139,6 +139,11 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`hooks-emit-prereq
+  <clang-tidy/checks/hooks-emit-prereq>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`hooks-entry-point-recursion
   <clang-tidy/checks/hooks-entry-point-recursion>` check.